SetPickupOptions changed to make the buttonAllocator to clear all buttons first
This fixes a bug caused by a vanilla quirk to call SetPickupOptions twice, once with the old options. On the first run the buttons are created and our hook creates them with the wrong optionMap. Because the buttonAllocator uses these buttons on the 2nd call, the correct values are never hooked to the buttons.  Fixes #19